+++ packages/components/src/components/button/button.tsx
@@ -40,7 +40,7 @@ export class Button {
   /** (optional) Button variant */
   @Prop() variant?: string = 'primary';
   /** (optional) If `true`, the button is disabled */
-  @Prop() disabled?: boolean = false;
+  @Prop({ reflect: true }) disabled?: boolean = false;
   /** (optional) Button type */
   @Prop() type?: 'reset' | 'submit' | 'button';
   /** (optional) The name of the button, submitted as a pair with the button's `value` as part of the form data */
@@ -101,6 +101,12 @@ export class Button {
         const fakeButton = document.createElement('button');
         if (this.type) {
           fakeButton.type = this.type;
+        }
+        if (this.value) {
+          fakeButton.value = this.value;
+        }
+        if (this.name) {
+          fakeButton.name = this.name;
         }
         fakeButton.style.display = 'none';
         parentForm.appendChild(fakeButton);