+++ packages/components/src/components/telekom/telekom-profile-menu/telekom-profile-menu.tsx
@@ -82,6 +82,7 @@ export class TelekomProfileMenu {
 
   @Prop() logoutLabel: string;
   @Prop() logoutUrl?: string;
+  @Prop() logoutHandler?: string;
 
   @State()
   menuOpen = false;
@@ -161,21 +162,25 @@ export class TelekomProfileMenu {
     );
   }
 
+  buildLogoutButton() {
+    return {
+      type: 'button',
+      name: this.logoutLabel || 'Logout',
+      href: this.logoutUrl || LOGOUT_DEFAULT,
+      variant: 'secondary',
+      onClick: this.logoutHandler,
+    };
+  }
+
   buildUserNavigation() {
     const divider = [{ type: 'divider' }];
 
     const userInfo = readData(this.userInfo);
-    if (!userInfo) {
-      // console.error("userInfo missing");
-    }
-    userInfo.type = 'userInfo';
-
-    let serviceLinks = readData(this.serviceLinks);
-    if (!serviceLinks) {
-      // console.error("serviceLinks missing");
-      serviceLinks = [];
+    if (userInfo) {
+      userInfo.type = 'userInfo';
     }
 
+    const serviceLinks = readData(this.serviceLinks) || [];
     for (const el of serviceLinks) {
       el.type = 'item';
     }
@@ -187,18 +192,13 @@ export class TelekomProfileMenu {
       icon: 'service-settings',
     };
 
-    const logout = {
-      type: 'button',
-      name: this.logoutLabel,
-      href: this.logoutUrl || LOGOUT_DEFAULT,
-      variant: 'secondary',
-    };
-
     let menu = [];
 
-    menu = menu.concat(userInfo);
+    if (userInfo) {
+      menu = menu.concat(userInfo);
+    }
 
-    if (!this.serviceLinksEmpty()) {
+    if (userInfo && !this.serviceLinksEmpty()) {
       menu = menu.concat(divider);
     }
 
@@ -211,13 +211,16 @@ export class TelekomProfileMenu {
       menu = menu.concat(divider);
     }
 
-    menu = menu.concat(logout);
+    menu = menu.concat(this.buildLogoutButton());
 
     return menu;
   }
 
   serviceLinksEmpty() {
-    return (this.hideLoginSettings && this.serviceLinks.length < 1) === true;
+    return (
+      this.hideLoginSettings &&
+      (!this.serviceLinks || this.serviceLinks.length < 1)
+    );
   }
 
   buildDesktopMenuStyles() {