+++ packages/storybook-vue/stories/components/checkbox/ScaleCheckbox.vue
@@ -11,6 +11,9 @@
     :input-id="inputId"
     :required="required"
     :aria-label-checkbox="ariaLabelCheckbox"
+    :hide-label="hideLabel"
+    :status="status"
+    :styles="styles"
     @scaleChange="scaleChange"
   >
   </scale-checkbox>
@@ -31,7 +34,11 @@ export default {
     value: { type: String },
     inputId: { type: String },
     required: { type: Boolean },
-    ariaLabelCheckbox: { type: String}
+    ariaLabelCheckbox: { type: String },
+    ariaDetailsId: { type: String },
+    hideLabel: { type: Boolean},
+    status: { type: String },
+    styles: { type: String },
   },
   methods: {
     scaleChange($event) {