+++ packages/storybook-vue/stories/components/radio-button-group/ScaleRadioButton.vue
@@ -10,15 +10,17 @@ export default {
     checked: { type: Boolean },
     value: { type: String },
     inputId: { type: String },
-    styles: { type: String }
+    styles: { type: String },
+    ariaDetailsId: { type: String },
+    status: { type: String , default: 'COMMENT: DEPRECATED - invalid should replace status'}
   },
   methods: {
     fireEvents() {
       this.$emit('scale-change');
-    }
+    },
   },
   render() {
     return this.$slots.default;
-  }
+  },
 };
 </script>