+++ packages/storybook-vue/stories/components/tab-navigation/TabHeader.vue
@@ -1,0 +1,11 @@
+<script>
+export default {
+  name: 'Tab Header',
+  props: {
+    styles: { type: String },
+    disabled: { type: Boolean, default: 'false' },
+    selected: { type: Boolean, default: 'false' },
+    size: { type: ['large', 'small'], default: 'small' },
+  },
+};
+</script>