+++ packages/components-angular/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-beta.155](https://github.com/telekom/scale/compare/v3.0.0-beta.154...v3.0.0-beta.155) (2024-11-26)
+
+**Note:** Version bump only for package @telekom/scale-components-angular
+
+
+
+
+
+
+# [3.0.0-beta.154](https://github.com/telekom/scale/compare/v3.0.0-beta.152...v3.0.0-beta.154) (2024-11-06)
+
+**Note:** Version bump only for package @telekom/scale-components-angular
+
+
+
+
+
+
+# [3.0.0-beta.153](https://github.com/telekom/scale/compare/v3.0.0-beta.152...v3.0.0-beta.153) (2024-09-19)
+
+**Note:** Version bump only for package @telekom/scale-components-angular
+
+
+
+
+
+
+# [3.0.0-beta.152](https://github.com/telekom/scale/compare/v3.0.0-beta.151...v3.0.0-beta.152) (2024-07-10)
+
+**Note:** Version bump only for package @telekom/scale-components-angular
+
+
+
+
+
+# [3.0.0-beta.151](https://github.com/telekom/scale/compare/v3.0.0-beta.150...v3.0.0-beta.151) (2024-04-12)
+
+**Note:** Version bump only for package @telekom/scale-components-angular
+
+
+
+
+
 # [3.0.0-beta.150](https://github.com/telekom/scale/compare/v3.0.0-beta.149...v3.0.0-beta.150) (2024-02-26)
 
 **Note:** Version bump only for package @telekom/scale-components-angular